Moved meetup API part to a new file
