+++ telegram-bot/main.go
@@ -1,12 +1,9 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"math/rand"
-	"net/http"
 	"os"
 	"strconv"
 	"time"
@@ -16,26 +13,6 @@ import (
 )
 
 var bot *tbot.BotAPI
-
-var finallist string //variable to store final list of meetups
-
-//-------structs to parse json data from Meetup API
-type venue struct {
-	Venue string `json:"name"`
-}
-type group struct {
-	GroupName string `json:"name"`
-}
-type meetuplist struct {
-	Name  string `json:"name"`
-	Venue venue  `json: "venue"`
-	Date  string `json:"local_date"`
-	Time  string `json:"local_time"`
-	Group group  `json: "group"`
-	Link  string `json:"link"`
-}
-
-//---------
 
 func start(ID int64) {
 	bot.Send(tbot.NewMessage(ID, "Hello , I'm OSDC Bot, Use /help to know more. To join OSDC group: https://t.me/jiitosdc"))
@@ -103,30 +80,12 @@ func help(ID int64) {
 	bot.Send(tbot.NewMessage(ID, msg))
 }
 
-//fetching details of meetup of the group's url
-func getmeetups(url string) {
-	resp, err := http.Get(url)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	var meetups []meetuplist
-	json.Unmarshal([]byte(body), &meetups)
-	if (len(meetups)) > 0 {
-		finallist = finallist + "Title -" + "\t" + meetups[0].Name + "\n" + "Date -" + "\t" + meetups[0].Date + "\n" + "Link -" + "\t" + meetups[0].Link + "\n\n"
-	}
-}
-
 //list of groups,keeps sending one by one to getmeetups()
 func ncrmeetups(ID int64) {
 	urlnames := []string{"ilugdelhi", "pydelhi", "GDGNewDelhi", "gdgcloudnd", "Paytm-Build-for-India", "jslovers", "Gurgaon-Go-Meetup", "Mozilla_Delhi", "PyDataDelhi", "React-Delhi-NCR", "OWASP-Delhi-NCR-Chapter"}
 	for _, each := range urlnames {
 		url := "http://api.meetup.com/" + each + "/events"
-		getmeetups(url)
+		getMeetups(url)
 	}
 	bot.Send(tbot.NewMessage(ID, finallist))
 }
@@ -151,6 +110,7 @@ func main() {
 	var err error
 	bot, err = tbot.NewBotAPI(os.Getenv("TELEGRAM_TOKEN"))
 	if err != nil {
+		fmt.Println("error in auth")
 		log.Panic(err)
 	}
